Notify user when password is changed
When a password change occurs we now notify the user by default. This is very
useful for detecting when something bad might be going on for a user. If a
user receives this email when they did not expect they can now talk to the
site owner to get the situation fixed.