Added login signals
This commits adds two new signals which are similiar to their sign up
counterparts. user_logged_in and user_login_attempt which are sent as they
are named.

Couple of other changes made to make the signals work such as unifying the
interface to get the identifier field for each login form and added
LoginView.after_login.