Simplified password reset conditions on email
This change allows any email address regardless of which state to be used for
password reset. The email sent is harmless as the action of password reset
is non-destructive.

This allows a simpler means for those without confirmed emails due to missing
the email or some other reason to reset their account password.