Blank lines should not contain any whitespace
Eldarion has decided for open source projects to conform to the blank lines
should not contain whitespace style guideline. While we still disagree with it,
we want to do less style guide enforcing with our contributors and make it
easier for them to contribute.

This does not mean we won't enforce this style guideline. Consistency is still
king and mixing whitespace lines with no whitespace lines is a big no-no. It
also still unacceptable to send us a pull request fixing two lines and
correcting whitespace in other parts of the code. Your pull requests should be
concise and not the product of a crazy editor that fixes all whitespace because
it thinks it is doing you a favor.