Write django-user-accounts instead of d-u-a
While reading the FAQ I stumbled over `d-u-a`. I hadn't seen it anywhere in the documentation. I searched the term before realizing it was `django-user-accounts`. (Duh.) I propose you spell it out and save the next person some confusion. Alternatively, consider referring to the project as `d-u-a` throughout for consistency.