Added migrations
Timezone choices from pytz has been vendored. The reason for this is
due to the way migrations are handled. A change in `choices` is
cause for a new migration. Since the timezones were programmatically
being set from pytz, it is possible for a user to install a newer
version and cause a migration.

South integration has been removed.