Extract Logic and Valiation from forms and views
So it will be possible to expose a REST interface the logic and validation
behind the forms and views needed to be extracted.  Two new files were
created, services.py and validators.py.

services.py:
Exports a static method 'signup' that contains the logic previously
contained in the signup view class.

Exports a static method update_email that contains the logic previously
contained in the settings view class.

validators.py:
Exports static methods that clean and validate username, email and passwords

The changes in forms.py and views.py is the removal of the above logic which
is replaced with calls to the newly exposed methods.